+++ src/main/java/org/sbrubbles/conditio/Condition.java
@@ -3,19 +3,18 @@ package org.sbrubbles.conditio;
 import org.sbrubbles.conditio.conditions.Signal;
 
 /**
- * Represents an unusual situation, which the running code doesn't know how to deal with. Conditions are meant to
- * be {@linkplain Scope#signal(Condition, Restart...) signalled}.
+ * Represents an unusual situation, which the running code doesn't know how to deal with, but the code that called it
+ * might. Conditions are meant to be {@linkplain Scope#signal(Condition, Restart...) signalled}, which is how
+ * lower-level code communicates what happened.
  * <p>
- * This interface provides a callback for {@code Scope.signal}, with a default implementation. This creates a
- * protocol which {@code Condition} subtypes may override. See {@link Signal} for an example.
- * <p>
- * In the default implementation, any subtypes of {@code Condition} which are not also subtypes of {@code Signal}
- * are <em>checked conditions</em>. This means that an exception will be thrown if no handler is found.
+ * This class is the superclass of all conditions in this library. It provides a callback for {@code Scope.signal},
+ * which creates a protocol that {@code Condition} subtypes may override. See {@link Signal} for an example.
+ * In the default implementation, an exception will be thrown if no handler is found.
  *
  * @see Scope#signal(Condition, Restart...)
  * @see Signal
  */
-public interface Condition {
+public class Condition {
   /**
    * Called when no handler was found.
    * <p>
@@ -26,7 +25,7 @@ public interface Condition {
    * @return the value to be returned by {@code signal}.
    * @throws HandlerNotFoundException if the implementation decides to error out.
    */
-  default Object onHandlerNotFound(Scope scope) throws HandlerNotFoundException {
+  public Object onHandlerNotFound(Scope scope) throws HandlerNotFoundException {
     throw new HandlerNotFoundException(this);
   }
 }