+++ src/main/java/org/sbrubbles/conditio/conditions/Signal.java
@@ -4,17 +4,19 @@ import org.sbrubbles.conditio.Condition;
 import org.sbrubbles.conditio.Scope;
 
 /**
- * {@code Signal} and its implementations are <em>unchecked conditions</em>, meaning that if no handler is found, they
+ * This class and its subclasses are <em>unchecked conditions</em>, meaning that if no handler is found, they
  * don't error out.
  * <p>
- * As a consequence, {@link Scope#signal signal}ing a {@code Signal} has no meaningful return value. {@code Signal}s
- * work better as notifications, which can be safely {@linkplain org.sbrubbles.conditio.restarts.Resume resumed}, and
- * maybe generate some useful side effects.
+ * Signals are meant for conditions that may go unhandled entirely. They work better as hints or notifications to
+ * higher-level code, which can be safely {@linkplain org.sbrubbles.conditio.restarts.Resume resumed}, and maybe
+ * generate some useful side effects.
+ * <p>
+ * As a consequence, {@link Scope#signal signal}ing a {@code Signal} has no meaningful return value.
  *
  * @see Condition
  * @see org.sbrubbles.conditio.restarts.Resume
  */
-public interface Signal extends Condition {
+public class Signal extends Condition {
   /**
    * This implementation does nothing.
    * <p>
@@ -23,7 +25,7 @@ public interface Signal extends Condition {
    * @return a garbage value, to be ignored.
    */
   @Override
-  default Void onHandlerNotFound(Scope scope) {
+  public Void onHandlerNotFound(Scope scope) {
     return null;
   }
 }