+++ src/main/java/org/sbrubbles/conditio/conditions/Warning.java
@@ -9,28 +9,44 @@ import java.io.PrintStream;
  *
  * @see org.sbrubbles.conditio.restarts.Resume
  */
-public interface Warning extends Signal {
+public class Warning extends Signal {
+  private final String message;
+  private final PrintStream output;
+
+  public Warning(String message) {
+    this(message, System.out);
+  }
+
+  public Warning(String message, PrintStream output) {
+    this.message = message;
+    this.output = output;
+  }
+
   /**
    * The message to print.
    *
    * @return the message to print.
    */
-  String getMessage();
+  public String getMessage() {
+    return message;
+  }
 
   /**
    * Where to print.
    *
    * @return where to print.
    */
-  PrintStream getOutput();
+  public PrintStream getOutput() {
+    return output;
+  }
 
   /**
    * Prints a message to the given print stream.
    */
   @Override
-  default Void onHandlerNotFound(Scope scope) {
+  public Void onHandlerNotFound(Scope scope) {
     getOutput().println("[WARN] " + getMessage());
 
-    return Signal.super.onHandlerNotFound(scope);
+    return super.onHandlerNotFound(scope);
   }
 }