+++ src/test/java/org/sbrubbles/conditio/fixtures/BasicCondition.java
@@ -4,7 +4,7 @@ import org.sbrubbles.conditio.Condition;
 
 import java.util.Objects;
 
-public class BasicCondition implements Condition {
+public class BasicCondition extends Condition {
   private final Object value;
 
   public BasicCondition(Object value) {