+++ src/test/java/org/sbrubbles/conditio/fixtures/PleaseSignalSomethingElse.java
@@ -2,4 +2,4 @@ package org.sbrubbles.conditio.fixtures;
 
 import org.sbrubbles.conditio.Condition;
 
-public class PleaseSignalSomethingElse implements Condition { /**/ }
+public class PleaseSignalSomethingElse extends Condition { /**/ }